Refactor: move handleTimeChange and handleDayOffChange to useTimeHandlers
