+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEvent, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 
 import { inject } from "@vercel/analytics";
 import "./App.css";
@@ -8,11 +8,12 @@ import HtmlToCanvas from "./components/HtmlToCanvas";
 import InfoAndReport from "./components/InfoAndReport";
 import Popup from "./components/Popup";
 import { tableHeaders, weekdays } from "./constants";
+import useTimeHandlers from "./hooks/useTimeHandlers";
 import { initialWorkTimesState, savedWorkTimeState, summaryTableState, workTimeState } from "./stores/atoms";
 import { overtimeStatusSelector, realWorkTimeMinutesSelector, remainingWorkTimeMinutesSelector } from "./stores/selectors";
 import { WeeklySummary } from "./types";
 import { saveLocalStorage } from "./utils/localStorageUtils";
-import { calculateRestTime, isValidWorkTime, minutesToTime } from "./utils/timeUtils";
+import { calculateRestTime, minutesToTime } from "./utils/timeUtils";
 
 inject();
 
@@ -28,45 +29,12 @@ const App: React.FC = () => {
     const [capturedImageURL, setCapturedImageURL] = useState("");
     const [showKakaoShareList, setShowKakaoShareList] = useState(false);
 
+    const { handleTimeChange, handleDayOffChange } = useTimeHandlers();
+
     useEffect(() => {
         saveLocalStorage("workTime", workTime);
         saveLocalStorage("savedWorkTime", savedWorkTime);
     }, [workTime, savedWorkTime]);
-
-    const handleTimeChange = (dayIndex: number, type: "start" | "end", event: ChangeEvent<HTMLInputElement>) => {
-        const inputValue = event.target.value;
-        const isValidValue = inputValue && /^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$/.test(inputValue);
-
-        const updatedWorkTime = workTime.map((dayItem, index) => {
-            if (index === dayIndex) {
-                const updatedDay = {
-                    ...dayItem,
-                    [type]: isValidValue ? inputValue : "",
-                };
-                const updatedTotal = isValidWorkTime({ start: updatedDay.start, end: updatedDay.end });
-                return { ...updatedDay, total: updatedTotal };
-            }
-            return dayItem;
-        });
-
-        setWorkTime(updatedWorkTime);
-    };
-
-    const handleDayOffChange = (type: "halfDay" | "fullDay", dayIndex: number, event: ChangeEvent<HTMLInputElement>) => {
-        const updatedWorkTime = workTime.map((dayItem, index) => {
-            if (index === dayIndex) {
-                const updatedDay = {
-                    ...dayItem,
-                    [type]: event.target.checked,
-                };
-                const updatedTotal = isValidWorkTime({ start: updatedDay.start, end: updatedDay.end });
-                return { ...updatedDay, total: updatedTotal };
-            }
-            return dayItem;
-        });
-
-        setWorkTime(updatedWorkTime);
-    };
 
     const handleClearInputs = () => {
         setWorkTime(initialWorkTimesState);
@@ -154,7 +122,7 @@ const App: React.FC = () => {
                                                 className="text-input"
                                                 placeholder="출근시간"
                                                 defaultValue={workTime[i] ? workTime[i].start : ""}
-                                                onChange={event => handleTimeChange(i, "start", event)}
+                                                onChange={event => handleTimeChange("start", i, event)}
                                             />
                                         )}
                                         {j === 2 && (
@@ -163,7 +131,7 @@ const App: React.FC = () => {
                                                 className="text-input"
                                                 placeholder="퇴근시간"
                                                 defaultValue={workTime[i] ? workTime[i].end : ""}
-                                                onChange={event => handleTimeChange(i, "end", event)}
+                                                onChange={event => handleTimeChange("end", i, event)}
                                             />
                                         )}
                                         {j === 3 ? (
